remove superseding support from generic-worker
