+++ workers/generic-worker/helper_test.go
@@ -11,7 +11,6 @@ import (
 	"io/ioutil"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -139,33 +138,6 @@ func testTask(t *testing.T) *tcqueue.TaskDefinitionRequest {
 		TaskGroupID:   taskGroupID,
 		WorkerType:    config.WorkerType,
 	}
-}
-
-func getArtifactContent(t *testing.T, taskID string, artifact string) ([]byte, *http.Response, *http.Response, *url.URL) {
-	queue := serviceFactory.Queue(config.Credentials(), config.RootURL)
-	url, err := queue.GetLatestArtifact_SignedURL(taskID, artifact, 10*time.Minute)
-	if err != nil {
-		t.Fatalf("Error trying to fetch artifacts from Amazon...\n%s", err)
-	}
-	t.Logf("Getting from url %v", url.String())
-	// need to do this so Content-Encoding header isn't swallowed by Go for test later on
-	tr := &http.Transport{
-		DisableCompression: true,
-	}
-	client := &http.Client{Transport: tr}
-	rawResp, _, err := httpbackoff.ClientGet(client, url.String())
-	if err != nil {
-		t.Fatalf("Error trying to fetch decompressed artifact from signed URL %s ...\n%s", url.String(), err)
-	}
-	resp, _, err := httpbackoff.Get(url.String())
-	if err != nil {
-		t.Fatalf("Error trying to fetch artifact from signed URL %s ...\n%s", url.String(), err)
-	}
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		t.Fatalf("Error trying to read response body of artifact from signed URL %s ...\n%s", url.String(), err)
-	}
-	return b, rawResp, resp, url
 }
 
 func ensureResolution(t *testing.T, taskID, state, reason string) {
